Change to Set_samplesToAverage causes TM corruption
The LPC TM does not decode properly. It's length is greater
than 8192, although the simulator does seem to pass it anyway.

Changing Set_samplesToAverage to 2 fixed the problem for now.

I suspect the issue is either in the logic which is evaluating
the frame counter, or in the hardwired array size of 300
for BinData[32][300] and HKData[16][300].
The TM message size was also greater than the max allowable
of 8k.